POG-105 [fix]: SummonerRecord Entity, BigInt -> Int로 변경
1. win, lose의 경우 BigInt에서 Int로 타입 변경

+++ libs/entity/src/domain/summonerRecord/SummonerRecord.entity.ts
@@ -26,12 +26,12 @@ export class SummonerRecord extends BaseTimeEntity {
   tier: string;
 
   @Column({
-    type: 'bigint',
+    type: 'int',
   })
   win: number;
 
   @Column({
-    type: 'bigint',
+    type: 'int',
   })
   lose: number;
